+++ src/main/java/ddingdong/ddingdongBE/domain/fileinformation/service/FileInformationService.java
@@ -5,8 +5,8 @@ import ddingdong.ddingdongBE.domain.fileinformation.entity.FileInformation;
 import ddingdong.ddingdongBE.domain.fileinformation.entity.FileTypeCategory;
 import ddingdong.ddingdongBE.domain.fileinformation.repository.FileInformationRepository;
 import ddingdong.ddingdongBE.file.FileStore;
-import ddingdong.ddingdongBE.file.dto.FileResponse;
-import ddingdong.ddingdongBE.file.dto.UploadFileDto;
+import ddingdong.ddingdongBE.file.service.dto.FileResponse;
+import ddingdong.ddingdongBE.file.service.dto.UploadFileDto;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
