+++ src/main/java/ddingdong/ddingdongBE/file/FileStore.java
@@ -1,10 +1,11 @@
 package ddingdong.ddingdongBE.file;
 
-import ddingdong.ddingdongBE.file.dto.UploadFileDto;
+import ddingdong.ddingdongBE.file.service.dto.UploadFileDto;
 
 import java.util.List;
 import org.springframework.web.multipart.MultipartFile;
 
+//TODO: 리팩토링 후 제거
 public interface FileStore {
 
     List<UploadFileDto> storeFile(List<MultipartFile> multipartFile, String fileType, String filePath);
